requirements.txt
fastapi
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
uvicorn
pydantic[email]
